pyproject.toml
[project]
name = "openai-agents-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chromadb>=1.1.0",
    "fastapi[all]>=0.117.1",
    "gradio>=4.44.1",
    "openai>=1.109.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]
